refactor: extract shared mobile backend wiring
Reuse existing persistence and sync modules while preserving Android behavior.

Refs #531
